+++ libs/ui/src/theme/index.ts
@@ -1,3 +1,4 @@
 export * from './baseTheme'
+export * from './chainAccentConfig'
 export * from './typings'
 export * from './ThemeColorVars'